requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
charset-normalizer==3.4.5
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.29
django-filter==25.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
kombu==5.6.2
openpyxl==3.1.5
packaging==26.0
pillow==11.3.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-decouple==3.8
reportlab==4.4.10
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
